getAbsolutePath() moved into utils.c
