+++ src/util.h
@@ -30,6 +30,8 @@
 #ifndef __REDIS_UTIL_H
 #define __REDIS_UTIL_H
 
+#include "sds.h"
+
 int stringmatchlen(const char *p, int plen, const char *s, int slen, int nocase);
 int stringmatch(const char *p, const char *s, int nocase);
 long long memtoll(const char *p, int *err);
@@ -37,5 +39,6 @@ int ll2string(char *s, size_t len, long long value);
 int string2ll(const char *s, size_t slen, long long *value);
 int string2l(const char *s, size_t slen, long *value);
 int d2string(char *buf, size_t len, double value);
+sds getAbsolutePath(char *filename);
 
 #endif
